+++ packages/shared/src/storage/GoogleDriveStorage/GoogleDriveStorage.ts
@@ -1,1 +1,4 @@
-export const googleDriveStorage = {};
+import type { GoogleDriveStorage } from './GoogleDriveStorage.android';
+
+export const googleDriveStorage: GoogleDriveStorage =
+  {} as unknown as GoogleDriveStorage;
